Implement FlexTabBar - highlight onClick. Reformat JS code.
