+++ test_runner_components/package.json
@@ -35,6 +35,8 @@
     "@plotly/dash-component-plugins": "^1.2.0",
     "babel-eslint": "^10.0.2",
     "babel-loader": "^8.0.6",
+    "babel-plugin-react-native-web": "0.17.1",
+    "bootstrap": "^5.1.0",
     "color": "^3.1.3",
     "copyfiles": "^2.1.1",
     "css-loader": "^3.0.0",
@@ -45,10 +47,9 @@
     "npm": "^6.1.0",
     "prop-types": "^15.7.2",
     "react": "17.0.1",
+    "react-bootstrap": "^1.6.1",
     "react-docgen": "^4.1.1",
     "react-dom": "17.0.1",
-    "react-native": "0.64.2",
-    "react-native-web": "0.17.1",
     "react-refresh": "^0.8.3",
     "react-scripts": "^4.0.3",
     "react-tabs": "^3.2.2",
