+++ test_runner_components/src/lib/components/ExampleArrowFuncComponentWithHook.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, {useState} from 'react';
 import PropTypes from "prop-types";
 
 /**
@@ -11,10 +11,10 @@ export const ExampleArrowFuncComponentWithHook = (props) => {
     const [count, setCount] = useState(0)
     return (
         <div>
-          <p>You clicked {count} times</p>
-          <button onClick={() => setCount(count + 1)}>
-            Click me
-          </button>
+            <p>You clicked {count} times</p>
+            <button onClick={() => setCount(count + 1)}>
+                Click me
+            </button>
         </div>
     )
 }
