+++ test_runner_components/src/lib/components/ExampleFuncComponentWithHook.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, {useState} from 'react';
 import PropTypes from 'prop-types';
 
 /**
@@ -8,16 +8,16 @@ import PropTypes from 'prop-types';
  * @constructor
  */
 export function ExampleFuncComponentWithHook(props) {
-  const [count, setCount] = useState(0)
+    const [count, setCount] = useState(0)
 
-  return (
-    <div>
-      <p>You clicked {count} times</p>
-      <button onClick={() => setCount(count + 1)}>
-        Click me
-      </button>
-    </div>
-  )
+    return (
+        <div>
+            <p>You clicked {count} times</p>
+            <button onClick={() => setCount(count + 1)}>
+                Click me
+            </button>
+        </div>
+    )
 }
 
 ExampleFuncComponentWithHook.propTypes = {
