+++ test_runner_components/src/lib/index.js
@@ -1,12 +1,12 @@
 /* eslint-disable import/prefer-default-export */
 import ExampleComponent from './components/ExampleComponent.react';
-import { ExampleFuncComponentWithHook } from './components/ExampleFuncComponentWithHook'
-import { ExampleArrowFuncComponentWithHook } from './components/ExampleArrowFuncComponentWithHook'
-import TabBar from './components/TabBar'
+import {ExampleFuncComponentWithHook} from './components/ExampleFuncComponentWithHook'
+import {ExampleArrowFuncComponentWithHook} from './components/ExampleArrowFuncComponentWithHook'
+import {FlexTabBar} from "./components/FlexTabBar";
 
 export {
     ExampleComponent,
     ExampleFuncComponentWithHook,
     ExampleArrowFuncComponentWithHook,
-    TabBar
+    FlexTabBar
 };
